fix: reorg misconduct file structure
