+++ srml/slashing/src/misconduct/mod.rs
@@ -1,0 +1,8 @@
+/// Babe misconduct types
+pub mod babe;
+/// GRANDPA misconduct types
+pub mod grandpa;
+/// Network misconduct types
+pub mod network;
+/// Example implementation of rolling misconduct
+pub mod example;
